backend/requirements.txt
# Django and Core Dependencies - Python 3.12 compatible
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# GraphQL
graphene-django==3.2.2
graphql-core==3.2.3

# Authentication & Security
PyJWT==2.8.0
cryptography==43.0.3
django-environ==0.11.2

# Development & Testing
pytest==7.4.4
pytest-django==4.8.0
factory-boy==3.3.0
coverage==7.4.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5